% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier}
\alias{classifier}
\title{Logarithm Marginal Density Ratios Transformation and Support Vector Machine}
\source{
For transformation, based on
Wang H , Gu J , Wang S . An effective intrusion detection framework based on SVM with feature augmentation[J]. Knowledge-Based Systems, 2017, 136(Nov.15):130-139.
Fan J , Feng Y , Jiang J , et al. Feature Augmentation via Nonparametrics and Selection (FANS) in High Dimensional Classification[J]. Journal of the American Statal Association, 2016, 111(513):275.
For SVM, based on
•	Chang, Chih-Chung and Lin, Chih-Jen:
  LIBSVM: a library for Support Vector Machines
  http://www.csie.ntu.edu.tw/~cjlin/libsvm
•	Exact formulations of models, algorithms, etc. can be found in the document:
  Chang, Chih-Chung and Lin, Chih-Jen:
  LIBSVM: a library for Support Vector Machines
  http://www.csie.ntu.edu.tw/~cjlin/papers/libsvm.ps.gz
•	More implementation details and speed benchmarks can be found on: Rong-En Fan and Pai-Hsune Chen and Chih-Jen Lin:
  Working Set Selection Using the Second Order Information for Training SVM
  http://www.csie.ntu.edu.tw/~cjlin/papers/quadworkset.pdf
}
\usage{
classifier(training, test, yname, levels, yflag = T, svm = c("svmt"))
}
\arguments{
\item{training}{original train set.}

\item{test}{original test set.}

\item{yname}{the name of result-column.}

\item{levels}{results in result-column.}

\item{yflag}{a flag that describes whether there is result-column in the test set.}

\item{svm}{type of support vector machine.}
}
\value{
The function gives the result of prediction base train and test set.
}
\description{
Train SVM model based on train set which is logarithm marginal density ratio transformed. Apply trained model on test set to predict.
}
\details{
First, apply logarithm marginal density ratio transformation on train and test set. Then, train SVM model based on transformed train set. Finally, exert the model to predict category by transformed test set.
}
\examples{
intest = subnum(rawdata, k = 10)[[1]]
test = rawdata[intest,]
training = rawdata[-intest,]
classifier(training, test, 'class', c('normal', 'attack'), T, c('scmf', 'svmt'))
}
