% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMDRT.R
\name{LMDRT}
\alias{LMDRT}
\title{Logarithm Marginal Density Ratios Transformation}
\source{
For transformation, based on
Wang H , Gu J , Wang S . An effective intrusion detection framework based on SVM with feature augmentation[J]. Knowledge-Based Systems, 2017, 136(Nov.15):130-139.
Fan J , Feng Y , Jiang J , et al. Feature Augmentation via Nonparametrics and Selection (FANS) in High Dimensional Classification[J]. Journal of the American Statal Association, 2016, 111(513):275.
}
\usage{
LMDRT(Train, Test, yname, levels, yflag)
}
\arguments{
\item{Train}{a train set that needs to be transformed.}

\item{Test}{a train set that needs to be transformed.}

\item{yname}{a train set that needs to be transformed.}

\item{levels}{results in result-column.}

\item{yflag}{a flag that describes whether there is result-column in the test set.}
}
\value{
The function gives transformed datasets from train and test set.
}
\description{
Features augmentation via logarithm marginal density ratios transformation in train and test set.
}
\details{
First, split train set into 2 part. Then, apply the first part on kernel estimation of class conditional densities and use the result of kernel density estimation to transform therefore augment features in train and test set.
}
\examples{
intest = subnum(rawdata, k = 10)[[1]]
test = rawdata[intest,]
training = rawdata[-intest,]
LMDRT(training, test, 'class', c('normal', 'attack'), T)
}
