% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logInPred.R
\name{logInPred}
\alias{logInPred}
\title{Feature Extraction}
\source{
For transformation, based on
Wang H , Gu J , Wang S . An effective intrusion detection framework based on SVM with feature augmentation[J]. Knowledge-Based Systems, 2017, 136(Nov.15):130-139.
Fan J , Feng Y , Jiang J , et al. Feature Augmentation via Nonparametrics and Selection (FANS) in High Dimensional Classification[J]. Journal of the American Statal Association, 2016, 111(513):275.
}
\usage{
logInPred(i, m, d0, d1, yname, Mkex = Mkex, Mkey = Mkey)
}
\arguments{
\item{i}{a index of feature that needs to be transformed.}

\item{m}{dataset consisting of features.}

\item{d0}{samples’ data in one class}

\item{d1}{samples’ data in the other class.}

\item{yname}{the name of result-column.}

\item{Mkex}{a function to estimate kernel density of x.}

\item{Mkey}{a function to estimate kernel density of y.}
}
\value{
The transform function outputs augmented feature.
}
\description{
Use the result of kernel density estimation to transform therefore augment certain feature of train/test set.
}
\details{
For dataset d0 and d1, the default algorithm use Mkex( ) and Mkey( ) to estimate kernel density. Use f( ) and g( ), the outcome of estimation, to transform the ith feature of dataset m by log(f(m)) - log(g(m)).
}
\examples{
intest = subnum(rawdata, k=10)[[1]]
test = rawdata[intest,]
training = rawdata[-intest,]
groups = subgroup(training, 'class', c('attack', 'normal'))
d0 = groups[[1]]
d1 = groups[[2]]
denx0 = Mkex(d0, yname)
deny0 = Mkey(d0, yname)
logInPred(1, training, d0, d1, class, Mkex = Mkex, Mkey = Mkey)
}
