% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup.R
\name{subgroup}
\alias{subgroup}
\title{Random Grouping II}
\source{
For sample, based on

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Ripley, B. D. (1987) Stochastic Simulation. Wiley.
}
\usage{
subgroup(training, yname, levels)
}
\arguments{
\item{training}{a train set that needs to be grouped by row.}

\item{yname}{the name of result-column.}

\item{levels}{results in result-column.}
}
\value{
The value returned is a list of 3 sets. The first set X1_0 and second set X1_1 are samples of two different levels in the first part of train set respectively. The third set X2 is the second part of train set.
}
\description{
Random grouping function for index of train set’s row to divide into 2 groups.
}
\details{
Randomly split train set into two mutually exclusive parts. Classify the first part based on levels of yname.
}
\examples{
intest = subnum(rawdata, k=10)[[1]]
test = rawdata[intest,]
training = rawdata[-intest,]
groups = subgroup(training, 'class', c('attack', 'normal'))
}
