% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnum.R
\name{subnum}
\alias{subnum}
\title{Random Grouping I}
\source{
For sample, based on

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Ripley, B. D. (1987) Stochastic Simulation. Wiley.
}
\usage{
subnum(data, k = 10)
}
\arguments{
\item{data}{dataset that needs to be grouped by row.}

\item{k}{the number of groups.}
}
\value{
The value returned is a list of k-group-index, each element of the list is the sample in indexes of rows of the grouping results.
}
\description{
Random grouping function for indexes of dataset’s row to divide into k groups.
}
\details{
Randomly split the number of dataset’s rows into k mutually exclusive groups.
Divide the number of dataset’s rows by k and round the outcome. The result is regarded as the number of samples of each group. Use the sample( ) to extract k-1 groups out in the index of dataset’s rows, until the last group remains.
}
\examples{
intest = subnum(rawdata, k=10)
}
