% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\title{Features Extraction}
\source{
For transformation, based on
Wang H , Gu J , Wang S . An effective intrusion detection framework based on SVM with feature augmentation[J]. Knowledge-Based Systems, 2017, 136(Nov.15):130-139.
Fan J , Feng Y , Jiang J , et al. Feature Augmentation via Nonparametrics and Selection (FANS) in High Dimensional Classification[J]. Journal of the American Statal Association, 2016, 111(513):275.
}
\usage{
trans(train, test, d0, d1, yname, yflag = T)
}
\arguments{
\item{train}{a train set that needs to be transformed.}

\item{test}{a test set that needs to be transformed.}

\item{d0}{samples’ data in one class}

\item{d1}{samples’ data in one class}

\item{yname}{the name of result-column.}

\item{yflag}{a flag that describes whether there is result-column in the test set.}
}
\value{
The transform function outputs dataset of augmented features.
}
\description{
Use the result of kernel density estimation to transform therefore augment features of train and test set.
}
\details{
For dataset d0 and d1, the default algorithm use Mkex( ) and Mkey( ) to estimate kernel density. Use f( ) and g( ), the outcome of estimation, to transform features in train and test set by log(f(x)) - log(g(x)).
}
\examples{
intest = subnum(rawdata, k=10)[[1]]
test = rawdata[intest,]
training = rawdata[-intest,]
groups = subgroup(training, 'class', c('attack', 'normal'))
d0 = groups[[1]]
d1 = groups[[2]]
train=groups[[3]]
data=trans(train,test,d0,d1,'class',yflag=T)
}
